chore: rename ui component files to kebab case
