+++ packages/ui-components/src/alert.tsx
@@ -9,7 +9,7 @@ import {
   TextStyle,
   PressableProps,
 } from 'react-native';
-import { useTheme } from './theme/ThemeProvider';
+import { useTheme } from './theme/theme-provider';
 
 // Context to share variant with subcomponents
 const AlertContext = createContext<{ variant: AlertVariant }>({
