+++ packages/ui-components/src/avatar.tsx
@@ -10,7 +10,7 @@ import {
   View,
   ViewStyle,
 } from 'react-native';
-import { useTheme } from './theme/ThemeProvider';
+import { useTheme } from './theme/theme-provider';
 
 interface AvatarContextValue {
   loaded: boolean;
