+++ packages/ui-components/src/bottom-navigation-item.tsx
@@ -6,7 +6,7 @@ import {
   StyleSheet,
   GestureResponderEvent,
 } from 'react-native';
-import { useTheme } from './theme/ThemeProvider';
+import { useTheme } from './theme/theme-provider';
 
 export interface BottomNavItemProps {
   icon: React.ReactNode;
