+++ packages/ui-components/src/bottom-navigation.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { BottomNavItem } from './bottom-navigation-item';
-import { Icon } from './theme/Icon';
-import { useTheme } from './theme/ThemeProvider';
+import { Icon } from './theme/icon';
+import { useTheme } from './theme/theme-provider';
 
 export interface BottomNavigationProps {
   activeTab?: string;
