+++ packages/ui-components/src/chat-item.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, Pressable } from 'react-native';
 import { Avatar, AvatarImage } from './avatar';
-import { useTheme } from './theme/ThemeProvider';
+import { useTheme } from './theme/theme-provider';
 
 export interface ChatItemProps {
   name: string;
