+++ packages/ui-components/src/message-bubble.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { MessageBubble, type MessageBubbleProps } from './message-bubble';
-import { ThemeProvider } from './theme/ThemeProvider';
+import { ThemeProvider } from './theme/theme-provider';
 
 type Scheme = 'light' | 'dark';
